requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
django-environ>=0.11,<0.12
psycopg2-binary>=2.9,<3.0
gunicorn>=21.2,<22.0
django-filter>=23.2
drf-spectacular>=0.27.0
flake8>=6.1.0
pytest-django>=4.5.2
